move marginals
