+++ src/components/marginals/index.js
@@ -1,0 +1,2 @@
+export { default as Footer } from "./Footer/Footer";
+export {default as Navbar } from "./Navbar/Navigation"
