+++ src/index.js
@@ -3,14 +3,14 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { Route, BrowserRouter as Router, Switch, Redirect } from 'react-router-dom';
-import Navigation from './components/Navigation';
+import Navigation from './components/marginals/Navbar/Navigation';
 import Home from './pages/Home';
 import Manifesto from './pages/Manifesto';
 import Journal from './pages/Journal';
 import Contact from './pages/Contact';
 import Login from './pages/Login';
 import SignUp from './pages/SignUp';
-import Footer from './components/Footer';
+import { Footer } from './components';
 import Header from './components/Header';
 
 function App() {
